pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mcp-llm"
version = "0.1.0"
description = "MCP Client CLI for interacting with Anthropic's Claude model and MCP servers"
authors = [
    {name = "Ling Li", email = "email@ling-li.com"},
]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "anthropic>=0.49.0",
    "asyncio>=3.4.3",
    "click>=8.1.7",
    "loguru>=0.7.3",
    "mcp>=0.0.17",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
]

[project.scripts]
mcpllm = "cli:app"

[tool.hatch.build.targets.wheel]
packages = ["mcp_llm"]
